Description:codex
Match-id-a8953a2fae193835b06fa3e731a53cba652e10f5

+++ common/include/input_hub.cpp
@@ -613,14 +613,14 @@ int32_t InputHub::QueryLocalTouchScreenInfo(int fd)
     info.localAbsInfo.absXMax = absInfo.maximum;
     info.localAbsInfo.absMtPositionXMin = absInfo.minimum;
     info.localAbsInfo.absMtPositionXMax = absInfo.maximum;
-    info.sinkPhyWidth = absInfo.maximum + 1;
+    info.sinkPhyWidth = (uint32_t)(absInfo.maximum + 1);
 
     ioctl(fd, EVIOCGABS(ABS_MT_POSITION_Y), &absInfo);
     info.localAbsInfo.absYMin = absInfo.minimum;
     info.localAbsInfo.absYMax = absInfo.maximum;
     info.localAbsInfo.absMtPositionYMin = absInfo.minimum;
     info.localAbsInfo.absMtPositionYMax = absInfo.maximum;
-    info.sinkPhyHeight = absInfo.maximum + 1;
+    info.sinkPhyHeight = (uint32_t)(absInfo.maximum + 1);
 
     ioctl(fd, EVIOCGABS(ABS_PRESSURE), &absInfo);
     info.localAbsInfo.absPressureMin = absInfo.minimum;
@@ -1009,12 +1009,12 @@ void InputHub::HandleTouchScreenEvent(struct input_event readBuffer[], const siz
         for (size_t j = iter.first; j <= iter.second; j++) {
             struct input_event &iev = readBuffer[j];
             if (iev.code == ABS_MT_POSITION_X) {
-                absX = iev.value;
-                absXIndex = j;
+                absX = (uint32_t)iev.value;
+                absXIndex = (int32_t)j;
             }
             if (iev.code == ABS_MT_POSITION_Y) {
-                absY = iev.value;
-                absYIndex = j;
+                absY = (uint32_t)iev.value;
+                absYIndex = (int32_t)j;
             }
         }
